libui: use uint64_t to avoid overflow issues with big buffers
Fixes add-overflow errors caught by ubsan when running AHardwareBuffer
VK tests.

Test: dEQP-VK.api.external.memory.android_hardware_buffer.image_formats#r16g16b16a16_sfloat

+++ libs/ui/GraphicBufferAllocator.cpp
@@ -72,7 +72,7 @@ size_t GraphicBufferAllocator::getTotalSize() const {
 void GraphicBufferAllocator::dump(std::string& result) const {
     Mutex::Autolock _l(sLock);
     KeyedVector<buffer_handle_t, alloc_rec_t>& list(sAllocList);
-    size_t total = 0;
+    uint64_t total = 0;
     result.append("Allocated buffers:\n");
     const size_t c = list.size();
     for (size_t i=0 ; i<c ; i++) {
